Move Controller and GuitarController into frc.robot.controller
